update aur_update.sh to v4.6-2

+++ .script/aur_update.sh
@@ -1,13 +1,17 @@
-###############################################################################################################
-# aur-update v4.5
-# script di aggiornamento pacchetti installati da AUR, by rebellion
-# preferibilmente da lanciare tramite terminale per visualizzare messaggi di dialogo con lo script
-# a es.: gnome-terminal --hide-menubar --title=\"AUR update\" -e \'sh -c \"sh /home/$USER/aur-update\"\'
-# opzione 'aur-update -p' per inserire il nome pacchetto da aggiornare
-# opzione 'aur-update -p $nomepacchetto' per aggiornare nomepacchetto in automatico
-# gli aggiornamenti sono organizzati in ~/.cache/AUR/$nomepacchetto dove sono presenti PKGBUILD, pkg compilato
-# e lo snapshot scaricato .tar.gz
-###############################################################################################################
+#!/bin/bash
+
+#############################################################################################################
+
+#aur-update v4.6-2
+#script di aggiornamento pacchetti installati da AUR, by rebellion
+#da lanciare tramite terminale per visualizzare messaggi di dialogo con lo script
+#a es.: gnome-terminal --geometry 75x25 --hide-menubar --title=\"AUR update\" -e \'sh -c \"sh /home/$USER/aur-update\"\'
+#opzione 'aur-update -p' per inserire il nome pacchetto da aggiornare
+#opzione 'aur-update -p $nomepacchetto' per aggiornare nomepacchetto in automatico
+#gli aggiornamenti sono organizzati in ~/.cache/AUR/$nomepacchetto dove sono presenti PKGBUILD, pkg compilato
+#e lo snapshot scaricato .tar.gz
+
+#############################################################################################################
 
 controllo_dipendenze () {
    check_dip_aur="ok" ; a="\"Depends\"" ; aa="\",\""
@@ -50,6 +54,7 @@ controllo_dipendenze () {
       echo ; echo "PKGBUILD in $package_dir/$package_name assente." ; read -s -p "[Premi invio per continuare]"
    fi
 }
+
 updating () {
       risultato=""
       for ((i=1;i<=$numero;i++)) ; do
@@ -77,26 +82,43 @@ updating () {
                echo -e "Aggiornamento "$package_name" in corso, non spengere il pc o la connessione.\nbackup "$package_dir_del"---> fatto\nbackup "$package_name_tar"---> fatto"
                if wget -P $package_dir https://aur.archlinux.org/cgit/aur.git/snapshot/$package_name_tar ; then
                   tar -xvzf $package_dir/$package_name_tar -C $package_dir
-                  cd "$package_dir"/"$package_name"
-                  if makepkg -s ; then
-                     echo "Disinstallazione $package_name" ; sudo pacman -R $package_name
-                     echo "Installazione $package_name scaricato" ; sudo pacman -U *.pkg.tar.xz
-                     [[ -d "$package_dir_name_old" ]] && echo -n "Rimozione $package_dir_name_old di backup..." && rm -r $package_dir_name_old && echo "fatto."
-                     package_pkg=$(find /home/$USER/.cache/AUR/$package_name -type f -name "$package_name*" | grep "pkg.tar.xz") ; mv $package_pkg $package_dir
-                     package_pkgbuild=$(find /home/$USER/.cache/AUR/$package_name -type f -name "$package_name*" | grep "PKGBUILD") ; mv $package_pkg $package_dir
-                     rm -r $package_dir_name ; mkdir $package_dir/$package_name
-                     package_pkg=$(find /home/$USER/.cache/AUR -type f -name "$package_name*" | grep "pkg.tar.xz") ; mv $package_pkg $package_dir_name
-                     package_pkgbuild=$(find /home/$USER/.cache/AUR -type f -name "$package_name*" | grep "PKGBUILD") ; mv $package_pkg $package_dir_name
-                     dialog --title "pacchetto $package_name" --backtitle "Aggiornamento AUR" --msgbox "Aggiornamento eseguito!" 7 60 ; clear ; risultato=$risultato" "$package_name"-->installato\n"
+                  read -p "Vuoi consultare il PKGBUILD scaricato? [s/n]" pkg
+                  case $pkg in
+                     "s")
+                        [[ -e "/usr/bin/nano" ]] && gnome-terminal --geometry 75x40 --hide-menubar --title "PKGBUILD" -e 'sh -c "nano /home/$USER/.cache/AUR/chrome-gnome-shell-git/PKGBUILD"' 2>/dev/null || gnome-terminal --geometry 50x5 --hide-menubar --title "PKGBUILD" -e 'sh -c "echo && echo \"Installa editor nano con ^sudo pacman -S nano^\" && read -sp \"[PREMI INVIO PER USCIRE]\""' 2>/dev/null
+                     ;;
+                  esac
+                  echo ; read -p "Confermi l'aggiornamento? [s/n]" pkg
+                  if [ "$pkg" = "s" ] ; then
+                     cd "$package_dir"/"$package_name"
+                     if makepkg -s ; then
+                        echo "Disinstallazione $package_name" ; sudo pacman -R $package_name
+                        echo "Installazione $package_name scaricato" ; sudo pacman -U *.pkg.tar.xz
+                        [[ -d "$package_dir_name_old" ]] && echo -n "Rimozione $package_dir_name_old di backup..." && rm -r $package_dir_name_old && echo "fatto."
+                        package_pkg=$(find /home/$USER/.cache/AUR/$package_name -type f -name "$package_name*" | grep "pkg.tar.xz") ; mv $package_pkg $package_dir
+                        package_pkgbuild=$(find /home/$USER/.cache/AUR/$package_name -type f -name "$package_name*" | grep "PKGBUILD") ; mv $package_pkgbuild $package_dir
+                        rm -r $package_dir_name ; mkdir $package_dir/$package_name
+                        package_pkg=$(find /home/$USER/.cache/AUR -type f -name "$package_name*" | grep "pkg.tar.xz") ; mv $package_pkg $package_dir_name
+                        package_pkgbuild=$(find /home/$USER/.cache/AUR -type f -name "$package_name*" | grep "PKGBUILD") ; mv $package_pkg $package_dir_name
+                        mv $package_dir/$package_name_tar $package_dir/$package_name
+                        dialog --title "pacchetto $package_name" --backtitle "Aggiornamento AUR" --msgbox "Aggiornamento eseguito!" 7 60 ; clear ; risultato=$risultato" "$package_name"-->installato\n"
+                     else
+                        echo "Aggiornamento fallito."
+                        echo -n "Rimozione cartella $package_dir_name (cartella scompattata $package_name_tar scaricato) e ripristino backup..."
+                        sudo rm -r $package_dir/$package_name
+                        [[ -f "$package_dir"/"$package_name_tar" ]] && rm $package_dir/$package_name_tar
+                        [[ -d "$package_dir_name_old" ]] && mv $package_dir_name_old $package_dir_name
+                        echo "fatto."
+                        read -sp "[Premi invio per continuare]"
+                        risultato=$risultato" "$package_name"-->non_installato_problemi_nella_compilazione(make)\n"
+                     fi
                   else
-                     echo "Aggiornamento fallito."
+                     risultato=$risultato" "$package_name"-->non_installato_per_scelta_utente\n"
                      echo -n "Rimozione cartella $package_dir_name (cartella scompattata $package_name_tar scaricato) e ripristino backup..."
                      sudo rm -r $package_dir/$package_name
                      [[ -f "$package_dir"/"$package_name_tar" ]] && rm $package_dir/$package_name_tar
                      [[ -d "$package_dir_name_old" ]] && mv $package_dir_name_old $package_dir_name
-                     echo "fatto."
-                     read -sp "[Premi invio per continuare]"
-                     risultato=$risultato" "$package_name"-->non_installato_problemi_nella_compilazione(make)\n"
+                     echo "fatto." ; sleep 2
                   fi
                else
                   [[ -d "$package_dir_name_old" ]] && mv $package_dir_name_old $package_dir_name
@@ -114,10 +136,12 @@ updating () {
       done
       clear ; echo "Riassunto aggiornamento:" ; echo -e "\n$risultato" ; echo ; echo "premi invio per terminare..." ; read -s -n 1
 }
+
 controllo_pacchetto () {
    quit=0 ; for i in $(pacman -Qqm) ; do if [ "$i" = "$AUR" ] ; then quit=1 ; fi ; done
    [[ $quit -eq 0 ]] && echo "pacchetto '$AUR' non installato." && exit 0
 }
+
 if [ "$1" = "-p" ] && [ "$2" = "" ] ; then clear ; read -p "Inserisci il pacchetto da aggiornare> " AUR ; numero_AUR=1 ; controllo_pacchetto
 elif [ "$1" != "-p" ] && [ "$1" != "" ] ; then echo "opzione '$1' non valida." ; exit 0
 elif [ "$1" = "-p" ] && [ "$2" != "" ] ; then AUR=$2 ; numero_AUR=1 ; controllo_pacchetto
@@ -136,6 +160,7 @@ if ping -q -c 1 -W 1 8.8.8.8 >/dev/null || ping -q -c 1 -W 1 google.com >/dev/nu
    aur=()
    aur_ok=()
    update=""
+   count=1
    for ((i=0;i<$numero_AUR;i++)) ; do ii=$((i+1)) ; aur[$i]=$(echo $AUR | awk -v x=$ii '{print$x}')
       tput cuu 1 ; tput cuf 47 ; echo "[$(($i+1))/$numero_AUR]" ; tput el ; echo ${aur[$i]} ; tput cuu 1
       versione_new=$(curl -s "https://aur.archlinux.org/rpc.php?v=5&type=info&arg=${aur[$i]}" | grep "Version") ; outofdate=$(echo $versione_new | grep "OutOfDate\":null")
@@ -148,7 +173,8 @@ if ping -q -c 1 -W 1 8.8.8.8 >/dev/null || ping -q -c 1 -W 1 google.com >/dev/nu
          if [ "$versione_new" != "$versione_old" ] && ([ "$hash_old" != "$hash_new" ] || [ "$hash_old" = "" ]) ; then
             if [ "$outofdate" != "" ] ; then descrizione="${LIGHT_GREEN}AGGIORNABILE${NC}" ; descr="AGGIORNABILE" ; else descrizione="${LIGHT_GREEN}AGGIORNABILE${NC}(${LIGHT_RED}OutOfDate${NC})" ; descr="AGGIORNABILE_(OutOfDate)" ; fi
             if [ "$hash_old" = "" ] ; then descrizione=$descrizione"_firme_non_confrontabili" ; fi
-            aur_ok+=(False ${aur[$i]} $versione_old $versione_new)
+            aur_ok+=($count ${aur[$i]} off)
+            count=$(($count+1))
             aur[$i]=${aur[$i]}" --->"$descrizione
          else if [ "$outofdate" != "" ] ; then descrizione="${LIGHT_WHITE}AGGIORNATO${NC}" ; else descrizione="${LIGHT_WHITE}AGGIORNATO${NC}(${LIGHT_RED}OutOfDate${NC})" ; fi ; aur[$i]=${aur[$i]}" --->"$descrizione
          fi
@@ -161,10 +187,12 @@ if ping -q -c 1 -W 1 8.8.8.8 >/dev/null || ping -q -c 1 -W 1 google.com >/dev/nu
       tput cuu 1 ; tput el ; echo "RISULTATO RICERCA PACCHETTI AGGIORNABILI:"
       for ((i=0;i<$numero_AUR;i++)) ; do echo -e ${aur[$i]} ; done ; echo -e "\tt=aggiorna tutto | s=seleziona aggiornamenti | altro=esci" ; read -n 1 -s azione
       if [ "$azione" = "s" ] ; then
-         update=$(zenity --list --title="PACCHETTI AUR AGGIORNABILI" --text="seleziona per aggiornare" --width=1000 --height=650 --checklist --separator=" " --column="" --column="PACCHETTO" --column="VERSIONE INSTALLATA" --column="VERSIONE DA INSTALLARE" "${aur_ok[@]}" 2>/dev/null) ; numero=$(($(grep -o " " <<<"$update" | wc -l)+1))
-          [[ "${update// /}" != "" ]] && updating
-      elif [ "$azione" = "t" ] ; then numero=$((${#aur_ok[@]}/4))
-         for ((i=0;i<$numero;i++)) ; do ii=$(( (i*4)+1 )) ; update=$update" "${aur_ok[$ii]} ; done
+         exec 3>&1 ; select=$(dialog --backtitle "Aggiornamento AUR" --checklist "Seleziona i pacchetti da aggiornare:" 20 70 ${#aur_ok[@]} ${aur_ok[@]} 2>&1 1>&3) ; exitcode=$? ; exec 3>&-
+         for i in $select ; do update=$update" "${aur_ok[$(((3*$i)-2))]} ; done
+         numero=$(grep -o " " <<<"$update" | wc -l)
+         [[ "${update// /}" != "" ]] && updating
+      elif [ "$azione" = "t" ] ; then numero=$((${#aur_ok[@]}/3))
+         for ((i=0;i<$numero;i++)) ; do ii=$(( ($i*3)+1 )) ; update=$update" "${aur_ok[$ii]} ; done
          updating
       fi
    fi
@@ -173,4 +201,6 @@ else
 fi
 else dialog --title "Controllo gestione pacchetti (PACMAN)" --backtitle "Aggiornamento AUR" --msgbox "C'è già in esecuzione il gestore pacchetti, attenderne la fine." 7 60
 fi
+
 exit 0
+
